requirements.txt
djangorestframework==3.15.1
django==4.2.11
pytest==8.1.1
requests==2.31.0